Fixed text alignment and scaling
