+++ src/WallpaperEngine/Render/CObject.h
@@ -13,6 +13,12 @@ class CScene;
 namespace WallpaperEngine::Render {
 class CObject : public Helpers::ContextAware, public TypeCaster {
 public:
+    struct ResolvedTransform {
+	glm::vec3 origin = glm::vec3 (0.0f);
+	glm::vec3 scale = glm::vec3 (1.0f);
+	float angle = 0.0f;
+    };
+
     CObject (Wallpapers::CScene& scene, const Object& object);
     virtual ~CObject () override = default;
 
@@ -23,6 +29,15 @@ public:
     [[nodiscard]] const AssetLocator& getAssetLocator () const;
     [[nodiscard]] int getId () const;
     [[nodiscard]] const Object& getObject () const;
+
+    [[nodiscard]] ResolvedTransform resolveTransform () const;
+    [[nodiscard]] ResolvedTransform resolveTransform (const WallpaperEngine::Data::Model::Object& object) const;
+
+    /**
+     * Computes the object's own transform (origin/scale/angle) without walking the
+     * parent chain. Used as the per-node step of resolveTransform.
+     */
+    [[nodiscard]] static ResolvedTransform localTransform (const WallpaperEngine::Data::Model::Object& object);
 
 private:
     Wallpapers::CScene& m_scene;
